[REFACTOR] Notification 서버 로직 분리 #82
