+++ umbba-notification/src/main/java/sopt/org/umbba/notification/config/sqs/SqsConfig.java
@@ -1,0 +1,34 @@
+package sopt.org.umbba.notification.config.sqs;
+
+import com.amazonaws.auth.AWSStaticCredentialsProvider;
+import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.services.sqs.AmazonSQSAsync;
+import com.amazonaws.services.sqs.AmazonSQSAsyncClient;
+import com.amazonaws.services.sqs.AmazonSQSAsyncClientBuilder;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+
+@Configuration
+public class SqsConfig {
+
+    @Value("${cloud.aws.credentials.accessKey}")
+    private String AWS_ACCESS_KEY;
+
+    @Value("${cloud.aws.credentials.secretKey}")
+    private String AWS_SECRET_KEY;
+
+    @Value("${cloud.aws.region.static}")
+    private String AWS_REGION;
+
+    @Primary
+    @Bean
+    public AmazonSQSAsync amazonSQSAsync() {
+        BasicAWSCredentials basicAWSCredentials = new BasicAWSCredentials(AWS_ACCESS_KEY, AWS_SECRET_KEY);
+        return AmazonSQSAsyncClientBuilder.standard()
+                .withRegion(AWS_REGION)
+                .withCredentials(new AWSStaticCredentialsProvider(basicAWSCredentials))
+                .build();
+    }
+}
