+++ umbba-notification/src/main/java/sopt/org/umbba/notification/controller/slack/SlackTestController.java
@@ -1,4 +1,4 @@
-package sopt.org.umbba.global.util.slack;
+package sopt.org.umbba.notification.controller.slack;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import sopt.org.umbba.global.common.dto.ApiResponse;
+import sopt.org.umbba.common.exception.dto.ApiResponse;
 
 @Slf4j
 @RestController
