+++ umbba-notification/src/main/resources/application.yaml
@@ -1,0 +1,73 @@
+cloud:
+  aws:
+    credentials:
+      accessKey: ${cloud-access}
+      secretKey: ${cloud-secret}
+    region:
+      static: ${cloud-region}
+    s3:
+      bucket: ${bucket-name}
+    stack:
+      auto: false
+    sqs:
+      notification:
+        name: ${AWS_SQS_NOTIFICATION_NAME_DEV}
+        url: ${AWS_SQS_NOTIFICATION_URL_DEV}
+      api:
+        name: ${AWS_SQS_API_NAME_DEV}
+        url: ${AWS_SQS_API_URL_DEV}
+
+
+spring:
+  datasource:
+    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: ${db-url}
+    username: ${db-user}
+    password: ${db-pwd}
+  jpa:
+    show-sql: false
+    hibernate:
+      ddl-auto: update
+      ejb:
+        naming_strategy: org.springframework.boot.orm.jpa.hibernate.SpringPhysicalNamingStrategy
+    properties:
+      hibernate:
+        format_sql: false
+
+kakao:
+  client-id: ${kakao-id}
+  authorization-grant-type: authorization_code
+  redirect-uri: ${kakao-redirect}
+
+apple:
+  iss: ${apple-iss}
+  client-id: ${apple-id}
+#  nonce: ${apple-nonce}
+
+jwt:
+  secret: ${jwt-secret}
+
+slack:
+  webhook:
+    url: ${slack-url}
+
+fcm:
+  key:
+    path: ${fcm-json-path}
+    scope: ${fcm-scope}
+  #    firebase-create-scoped: "https://www.googleapis.com/auth/firebase.messaging"
+  api:
+    url: ${fcm-api-url}
+  topic:
+    "qna_notification"
+
+logging:
+  level:
+    org:
+      hibernate:
+        type:
+          descriptor:
+            sql: debug
+
+server:
+  port: 9091
